// apps/breakout-game/src/types/PowerUpType.ts
export enum PowerUpType {
  MULTI_BALL = 'multiball',
  PADDLE_GROW = 'expand',
  PADDLE_SHRINK = 'shrink',
  SPEED_DOWN = 'slow',
  SPEED_UP = 'fast',
  EXTRA_LIFE = 'extraLife',
  LASER = 'laser',
  STICKY = 'sticky',
  SHIELD = 'shield',
  FIREBALL = 'fireball',
  SCORE_MULTIPLIER = 'scoreMultiplier'
}